feat(metronome): remove `lower` value
It doesn't combine well with subdivision
