+++ lib/metronome/metronome_screen.dart
@@ -6,7 +6,7 @@ import 'package:musbx/metronome/bpm_tapper.dart';
 import 'package:musbx/metronome/play_button.dart';
 import 'package:musbx/metronome/metronome.dart';
 import 'package:musbx/metronome/subdivisions.dart';
-import 'package:musbx/metronome/time_signature.dart';
+import 'package:musbx/metronome/higher.dart';
 import 'package:musbx/screen/card_list.dart';
 import 'package:musbx/screen/default_app_bar.dart';
 
@@ -29,33 +29,32 @@ Change the sound of beats by tapping them. Long press to remove and plus to add.
       ),
       body: CardList(
         children: [
+          Center(
+            child: PlayButton(
+              size: 150,
+            ),
+          ),
           Column(children: [
             Row(
               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               children: [
-                BpmButtons(
-                  iconSize: 50,
-                  fontSize: 45,
+                Expanded(
+                  child: BpmButtons(
+                    iconSize: 50,
+                    fontSize: 45,
+                  ),
                 ),
                 BpmTapper()
               ],
             ),
             BpmSlider(),
           ]),
-          Column(
-            children: [
-              TimeSignature(),
-              SizedBox(height: 8.0),
-              Subdivisions(),
-            ],
-          ),
           Column(children: [
-            Center(
-              child: PlayButton(
-                size: 150,
-              ),
-            ),
+            Higher(),
+            SizedBox(height: 8.0),
             CountDisplay(),
+            SizedBox(height: 16.0),
+            Subdivisions(),
           ]),
         ],
       ),
